perf: cut the zkVM guest jumpdest scan to four instructions per byte (#13216)
* perf: cut the zkVM guest jumpdest scan to four instructions per byte

The guest reaches every code byte through the scalar byte scanner: the word
scanner's sampling gate sent every contract analysed over a mainnet block down
the byte path anyway, and it cannot win when all real bytecode is PUSH-dense.
Drop the gate and cheapen the byte path instead.

Loading with lb rather than lbu sign-extends everything above PUSH32 to below
JUMPDEST, so one signed compare replaces the rebase-and-range-test. The walk
moves a ref byte instead of recomputing base+index, takes a single advance no
path branches over, and reads its thresholds from a static readonly int[] --
ILC re-materialises a compared-against constant at every use inside a loop and
folds static readonly scalars straight back into that, while an array element
stays in a register.

The plain-byte path is now lb, blt, addi, bltu. 9.93 -> 5.70 steps per visited
byte (-42.6%); guest cost -1.268%. Host codegen is unchanged.

* refactor: split the jumpdest scan into std and zkevm files

The two `#if ZK_EVM` regions become a `JumpDestinationAnalyzer.std.cs` /
`.zkevm.cs` pair, which is how the rest of the repo carries a per-flavour
body. The word scanner and its helpers now leave the guest compilation
structurally rather than by relying on ILC to trim them.

Also document the two invariants the review asked for: the positional
contract of `_byteScanThresholds`, and that a tail PUSH can leave the
moving reference past the end of the buffer, where it is only compared.

* fix: pin the guest jumpdest walk instead of overshooting a managed ref

A PUSH near the tail advanced the walk up to 32 bytes past the code, which as a
`ref byte` is a managed pointer outside its object - invalid for a relocating GC
even when it is only compared, never dereferenced. The zkVM guest's GC does not
relocate, but GuestJumpDestinationTests runs the same scan on CoreCLR.

Pinning and stepping a raw pointer removes the assumption. The per-byte loop is
unchanged: JIT disasm before and after is instruction-for-instruction identical,
with the pinned slot's two stores outside the loop.

* docs: point the std jumpdest scan at its zkevm sibling

Matches the convention the other .std.cs/.zkevm.cs pairs carry, and names the
differential test that keeps the two in step.
