+++ src/Nethermind/Nethermind.Evm.ZkEvm.Test/GuestJumpDestinationTests.cs
@@ -11,13 +11,13 @@ namespace Nethermind.Evm.ZkEvm.Test;
 
 /// <summary>Differential tests for <see cref="JumpDestinationAnalyzer"/>'s scalar scan.</summary>
 /// <remarks>
-/// The scan is word-at-a-time SWAR, and it is what a guest build runs: everywhere else
-/// <c>CreateJumpDestinationBitmap</c> finds <c>Vector512</c> or <c>Vector128</c> accelerated and takes
-/// one of those instead, so the scalar form is never reached through the real entry point. The bitmap
-/// feeds <see cref="JumpDestinationAnalyzer.ValidateJump"/>, so a wrong bit is wrong execution rather
-/// than a slowdown. It is compared against the obvious byte-at-a-time reference over the shapes where a
-/// word-wise walk could diverge from it: PUSH data truncated by the end of the code, the 64-bit bitmap
-/// segment boundary the flags are flushed on, every opcode value, and random bytecode.
+/// The scalar scan is what a guest build runs: everywhere else <c>CreateJumpDestinationBitmap</c> finds
+/// <c>Vector512</c> or <c>Vector128</c> accelerated and takes one of those instead, so it is never reached
+/// through the real entry point. The bitmap feeds <see cref="JumpDestinationAnalyzer.ValidateJump"/>, so a
+/// wrong bit is wrong execution rather than a slowdown. The guest walks a moving pointer and accumulates
+/// flags per 64-bit segment; it is compared against the obvious indexed reference over the shapes where the
+/// two could diverge: PUSH data truncated by the end of the code, the segment boundary the flags are
+/// flushed on, every opcode value, and random bytecode.
 /// </remarks>
 public class GuestJumpDestinationTests
 {
@@ -56,8 +56,8 @@ public class GuestJumpDestinationTests
     public void Scan_matches_the_reference(byte[] code) => AssertMatchesReference(code);
 
     /// <remarks>
-    /// The scan classifies a byte by unsigned range checks against <c>[JUMPDEST, PUSH32]</c>, so it is
-    /// only sound across the whole byte range - which this walks. The trailing JUMPDESTs are enough that
+    /// The scan classifies a byte by comparing it <em>signed</em> against <c>[JUMPDEST, PUSH32]</c>, which
+    /// is only equivalent across the whole byte range - which this walks. The trailing JUMPDESTs are enough that
     /// even a PUSH32 immediate stays inside the code, so the bytes an opcode masks are visible rather
     /// than truncated.
     /// </remarks>
